# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for ingesting, storing, and querying log data with Grok pattern support",
        "title": "LogSonic API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/ingest/end": {
            "post": {
                "description": "End the specified log ingest session and cleanup its resources",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingest"
                ],
                "summary": "End log ingest session",
                "parameters": [
                    {
                        "description": "Session end request with session_id",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.IngestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.IngestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/ingest/logs": {
            "post": {
                "description": "Ingest log data using existing Grok patterns and store them into the index",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingest"
                ],
                "summary": "Ingest log data",
                "parameters": [
                    {
                        "description": "Log ingest request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.IngestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.IngestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ingest/start": {
            "post": {
                "description": "Start a new log ingest session with specific options and returns a session ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingest"
                ],
                "summary": "Start log ingest session",
                "parameters": [
                    {
                        "description": "Log ingest session start request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.IngestSessionOptions"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.IngestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "timeresolve.Anchor": {
            "type": "object",
            "properties": {
                "kind": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "timeresolve.Resolution": {
            "type": "object",
            "properties": {
                "anchor": {
                    "$ref": "#/definitions/timeresolve.Anchor"
                },
                "force_mode": {
                    "type": "string"
                },
                "forced_day": {
                    "type": "integer"
                },
                "forced_month": {
                    "type": "integer"
                },
                "forced_year": {
                    "type": "integer"
                },
                "rollover": {
                    "type": "boolean"
                },
                "source_field": {
                    "description": "SourceField names a non-canonical capture to use as the line's\ntimestamp string. When set, the resolver reads fields[SourceField]\ninstead of probing the canonical names (timestamp/date/time/...).\nEmpty means \"use canonical scan\" (the legacy default).",
                    "type": "string"
                },
                "source_format": {
                    "description": "SourceFormat hints how to parse SourceField's value. Empty means\nauto-detect via dateparse + unix-epoch length heuristics. Special\nvalues: \"unix_seconds\" / \"unix_millis\" / \"unix_nanos\". Anything\nelse is treated as a Go time layout (e.g. \"2006-01-02-15.04.05.000000\"\nfor the BGL supercomputer format).",
                    "type": "string"
                },
                "timezone": {
                    "$ref": "#/definitions/timeresolve.TimezoneCfg"
                },
                "year_strategy": {
                    "type": "string"
                }
            }
        },
        "timeresolve.TimezoneCfg": {
            "type": "object",
            "properties": {
                "kind": {
                    "type": "string"
                },
                "value": {
                    "description": "IANA zone name when Kind=forced",
                    "type": "string"
                }
            }
        },
...
        "types.ErrorResponse": {
            "description": "Standardized error response structure used across all API endpoints",
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code for programmatic handling",
                    "type": "string"
                },
                "details": {
                    "description": "Additional error details",
                    "type": "string"
                },
                "error": {
                    "description": "Main error message",
                    "type": "string"
                },
                "status": {
                    "description": "Status will always be \"error\" for error responses",
                    "type": "string"
                }
            }
        },
...
        "types.IngestRequest": {
            "type": "object",
            "properties": {
                "logs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "session_id": {
                    "type": "string"
                }
            }
        },
        "types.IngestResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "failed": {
                    "type": "integer"
                },
                "filename": {
                    "type": "string"
                },
                "processed": {
                    "type": "integer"
                },
                "session_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "types.IngestSessionOptions": {
            "type": "object",
            "properties": {
                "custom_patterns": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "force_start_day": {
                    "type": "string"
                },
                "force_start_month": {
                    "type": "string"
                },
                "force_start_year": {
                    "type": "string"
                },
                "force_timezone": {
                    "type": "string"
                },
                "meta": {
                    "description": "Meta contains additional fields to be added to each log entry.\nThese fields are merged directly into the JSON output for each log.",
                    "type": "object",
                    "additionalProperties": true
                },
                "multiline": {
                    "description": "Multiline configures folding of physical lines into logical log\nrecords before pattern matching, for formats where a single log\nstatement spans multiple lines (stack traces, multi-line JSON,\netc). Nil/disabled preserves the historical one-line-per-record\nbehaviour. Folding is applied on /parse (including autosuggest)\nas well as ingest and live tail.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.MultilineConfig"
                        }
                    ]
                },
                "name": {
                    "type": "string"
                },
                "pattern": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "smart_decoder": {
                    "type": "boolean"
                },
                "source": {
                    "type": "string"
                },
                "source_mtime": {
                    "description": "SourceMTime is the modification time of the source file when\nknown. Lets the resolver anchor year-less or 2-digit-year\ntimestamps against the file rather than wall-clock now.",
                    "type": "string"
                },
                "timestamp_config": {
                    "description": "TimestampConfig is the user-confirmed resolution from the import\nwizard. When nil, the resolver auto-derives defaults from a\nsample (legacy ForceStart* fields are still honoured).",
                    "allOf": [
                        {
                            "$ref": "#/definitions/timeresolve.Resolution"
                        }
                    ]
                }
            }
        },
...
        "types.MultilineConfig": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "header_pattern": {
                    "description": "HeaderPattern is a regular expression; required when Mode==\"header\".\nA line matching it starts a new record, everything else is folded\ninto the preceding record. Continuation lines are joined with\nspaces (log2grok's JoinMultilineStrings behaviour).",
                    "type": "string"
                },
                "max_bytes": {
                    "type": "integer"
                },
                "max_lines": {
                    "description": "MaxLines caps continuation lines folded into one record (default\n100). MaxBytes caps the folded record's byte length (default 1MiB).",
                    "type": "integer"
                },
                "mode": {
                    "description": "\"header\" | \"indent\"",
                    "type": "string"
                }
            }
        },
...
    }
}